Performance improvements and new sorting mechanism.
Linux and Darwin based builds now have a significant improvement in the extraction of processes information by replacing the usage of the gopsutil/v3/process package with the execution of the ps command. Windows builds use the aforementioned package. The new sorting mecanism comes in Evertras/bubble-table==0.14.6
