+++ stats.go
@@ -4,7 +4,6 @@ import (
 	"github.com/shirou/gopsutil/v3/cpu"
 	"github.com/shirou/gopsutil/v3/disk"
 	"github.com/shirou/gopsutil/v3/mem"
-	"github.com/shirou/gopsutil/v3/process"
 )
 
 // TODO: Determine which constant better suits a given value.
@@ -102,32 +101,4 @@ func extractDiskInfo() []diskInfo {
 	}
 
 	return disks
-}
-
-func extractProcessesInfo() []processInfo {
-	ps, _ := process.Processes()
-	var processes []processInfo
-
-	for _, p := range ps {
-		u, _ := p.Username()
-		n, _ := p.Name()
-		prio, _ := p.Nice()
-		cPcg, _ := p.CPUPercent()
-		exeP, _ := p.Exe()
-		cmdL, _ := p.Cmdline()
-
-		processInfo := processInfo{
-			PId:           p.Pid,
-			User:          u,
-			Name:          n,
-			Priority:      prio,
-			CpuPercentage: cPcg,
-			Cmdline:       cmdL,
-			ExeP:          exeP,
-		}
-
-		processes = append(processes, processInfo)
-	}
-
-	return processes
 }
